Conflicts:
	src/condor_gridmanager/amazonjob.cpp
	src/condor_gridmanager/amazonresource.cpp
	src/condor_gridmanager/condorjob.cpp
	src/condor_gridmanager/creamjob.cpp
	src/condor_gridmanager/gahp-client.cpp
	src/condor_gridmanager/globusjob.cpp
	src/condor_gridmanager/gt4job.cpp
	src/condor_gridmanager/nordugridjob.cpp
	src/condor_gridmanager/nordugridresource.cpp
	src/condor_gridmanager/unicorejob.cpp

+++ src/condor_startd.V6/ResAttributes.h
@@ -224,9 +224,9 @@ private:
 
 		// These hold the fractions of shared, dynamic resources
 		// that are allocated to this CPU.
-	float			c_virt_mem_fraction;
+	double			c_virt_mem_fraction;
 
-	float			c_disk_fraction; // share of execute dir partition
+	double			c_disk_fraction; // share of execute dir partition
 	MyString        c_execute_dir;
 	MyString        c_execute_partition_id;  // unique id for partition
